+++ src/main/java/com/cinchapi/ccl/grammar/FunctionTokenSymbol.java
@@ -18,25 +18,14 @@ package com.cinchapi.ccl.grammar;
 import com.cinchapi.ccl.type.Function;
 
 /**
- * A {@link ValueTokenSymbol} that represents a {@link Function}.
- *
- * @author Jeff Nelson
+ * A {@link FunctionTokenSymbol} that contains a {@link Function}.
  */
-public class FunctionValueSymbol extends ValueTokenSymbol<Function>
-        implements FunctionTokenSymbol {
+public interface FunctionTokenSymbol extends Symbol {
 
     /**
-     * Construct a new instance.
+     * Return the {@link Function} associated with this symbol.
      * 
-     * @param value
+     * @return the {@link Function}
      */
-    public FunctionValueSymbol(Function value) {
-        super(value);
-    }
-
-    @Override
-    public Function function() {
-        return this.value;
-    }
-
+    public Function function();
 }